requirements.txt
dropbox==9.4.0
mss==4.0.3
tweepy==3.8.0
cryptography==2.6.1
stepic==0.4.1
pyfiglet==0.8.post1
requests==2.21.0
beautifulsoup4==4.8.2
Pillow==6.2.1
termcolor==1.1.0
